separated filter module from module responsible for building frames
